// src/api-client/models/ServiceTypeModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AquaTrail_Data
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ServiceTypeModel
 */
export interface ServiceTypeModel {
    /**
     * 
     * @type {string}
     * @memberof ServiceTypeModel
     */
    key?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceTypeModel
     */
    value?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceTypeModel
     */
    iconKey?: string | null;
}

/**
 * Check if a given object implements the ServiceTypeModel interface.
 */
export function instanceOfServiceTypeModel(value: object): value is ServiceTypeModel {
    return true;
}

export function ServiceTypeModelFromJSON(json: any): ServiceTypeModel {
    return ServiceTypeModelFromJSONTyped(json, false);
}

export function ServiceTypeModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): ServiceTypeModel {
    if (json == null) {
        return json;
    }
    return {
        
        'key': json['key'] == null ? undefined : json['key'],
        'value': json['value'] == null ? undefined : json['value'],
        'iconKey': json['iconKey'] == null ? undefined : json['iconKey'],
    };
}

export function ServiceTypeModelToJSON(value?: ServiceTypeModel | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'key': value['key'],
        'value': value['value'],
        'iconKey': value['iconKey'],
    };
}